perf(h264): optimize MC, thread pool, and decode hot paths
Major motion compensation refactor and performance improvements across
the H.264 decode pipeline.

MC (mc.rs):
- Pre-allocated bi-prediction scratch buffers (eliminate per-call alloc)
- Inline hints on hot filter functions
- Consolidated luma/chroma MC dispatch
- Edge-clamped reference block extraction with padding

Thread pool (thread_pool.rs):
- Postponed deblocking model (FFmpeg's postpone_filter pattern)
- Serial deblock pass after parallel slice decode
- Work queue improvements for multi-slice frames

Decode pipeline:
- CABAC: inline hints on hot context model functions
- CAVLC: expanded coefficient token tables for direct lookup
- Deblock: filter strength cache, reduced redundant neighbor checks
- IDCT: inline hints on butterfly operations
- MB: bi-prediction scratch buffers in FrameDecodeContext
- DPB: simplified reference frame management

+++ codecs/wedeo-codec-h264/src/tables.rs
@@ -44,7 +44,7 @@ pub const FIELD_SCAN_8X8: [u8; 64] = [
 ///
 /// From FFmpeg `zigzag_scan8x8_cavlc` in h264_slice.c (non-transposed variant).
 /// Formula: `zigzag_scan8x8_cavlc[i] = zigzag_scan8x8[(i/4) + 16*(i%4)]`.
-pub const ZIGZAG_SCAN_8X8_CAVLC: [[usize; 16]; 4] = [
+pub const ZIGZAG_SCAN_8X8_CAVLC: [[u8; 16]; 4] = [
     // Sub-scan 0 (indices 0-15)
     [0, 9, 17, 18, 12, 40, 27, 7, 35, 57, 29, 30, 58, 38, 53, 47],
     // Sub-scan 1 (indices 16-31)
@@ -430,8 +430,8 @@ mod tests {
         for sub in &ZIGZAG_SCAN_8X8_CAVLC {
             for &pos in sub {
                 assert!(pos < 64, "position {pos} out of range");
-                assert!(!seen[pos], "position {pos} appears twice");
-                seen[pos] = true;
+                assert!(!seen[pos as usize], "position {pos} appears twice");
+                seen[pos as usize] = true;
             }
         }
         assert!(seen.iter().all(|&s| s), "not all positions covered");
